Usunieta czesc warningow, interfejsowanie NewChart, zmieniony filtr EMG
git-svn-id: https://localhost:8443/svn/hm-mde-1@694 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ trunk/src/edr/src/hmm_view/DataFilterWidget.cpp
@@ -96,7 +96,7 @@ void DataFilterWidget::setColor( const QColor& color )
 
 const FilterEntryWidget* DataFilterWidget::getEntry( int index ) const
 {
-    UTILS_ASSERT(index >= 0 && index < entries.size());
+    UTILS_ASSERT(index >= 0 && index < static_cast<int>(entries.size()));
     return entries[index];
 }
